+++ lib/lib.rs
@@ -9,7 +9,7 @@ Just add the following to your `Cargo.toml`:
 
 ```toml
 [dependencies]
-afire = "0.2.0*"
+afire = "0.2.0"
 ```
 
 ## 📄 Info
@@ -61,7 +61,7 @@ For these you will need to enable the feature.
 To use these extra features enable them like this:
 
 ```toml
-afire = { version = "0.1.6", features = ["rate_limit", "logging"] }
+afire = { version = "0.2.0", features = ["rate_limit", "logging"] }
 ```
 
 - Threading